refactor: Burndown 차트 및 Effort Metrics 계산 로직을 별도의 훅으로 분리 (CM-205)
